tts-service/requirements.txt
grpcio
protobuf
piper-tts